@@ -18,19 +18,19 @@
   "devDependencies": {
     "@types/chai": "^4.2.22",
     "@types/mocha": "^9.0.0",
-    "@typescript-eslint/eslint-plugin": "^5.2.0",
-    "@typescript-eslint/parser": "^5.2.0",
+    "@typescript-eslint/eslint-plugin": "^5.3.1",
+    "@typescript-eslint/parser": "^5.3.1",
     "chai": "^4.3.4",
-    "eslint": "^8.1.0",
+    "eslint": "^8.2.0",
     "mocha": "^9.1.3",
     "ts-loader": "^9.2.6",
     "ts-node": "^10.4.0",
     "typescript": "^4.4.4",
-    "webpack": "^5.61.0",
+    "webpack": "^5.64.0",
     "webpack-bundle-analyzer": "^4.5.0",
     "webpack-cli": "^4.9.1"
   },
   "dependencies": {
-    "@types/node": "^16.11.6"
+    "@types/node": "^16.11.7"
   }
 }
